requirements.txt
python-nmap>=0.7.1
requests>=2.28.1
colorama>=0.4.6
msgpack>=1.0.4